Able to add 1 recipe to HomeScreen
Since we are doing new HomeScreen() every time we add a recipe then we are working with a new one every time we make a new recipe. That's why we can only add at most 1 recipe. However, this code can probably be reused when we implement the server. In the sense that when we create a new recipe we will push it to the server. Then within that scope we will create a new HomeScreen() and then get from the server the list of recipes that have been pushed so far. We would then iterate that list of recipes and add them to the RecipeList of the homescreen